refactor: shared fileSystem utils
