+++ electron/app/extensions/downloadExtensionEntry.ts
@@ -2,7 +2,14 @@ import log from 'loglevel';
 import fetch from 'node-fetch';
 import path from 'path';
 
-import {createFolder, createOrRecreateFolder, deleteFile, doesPathExist, writeFile} from '../utils/extensions';
+import {
+  createFolder,
+  createOrRecreateFolder,
+  deleteFile,
+  doesPathExist,
+  writeFile,
+} from '@monokle-desktop/shared/utils/fileSystem';
+
 import {DownloadExtensionEntryOptions} from './types';
 
 async function downloadExtensionEntry<ExtensionEntryType>(
