+++ electron/app/extensions/loadExtension.ts
@@ -2,8 +2,8 @@ import log from 'loglevel';
 import path from 'path';
 
 import type {AnyExtension} from '@monokle-desktop/shared/models';
+import {doesPathExist, readFile} from '@monokle-desktop/shared/utils/fileSystem';
 
-import {doesPathExist, readFile} from '../utils/extensions';
 import {LoadExtensionOptions} from './types';
 
 async function loadExtension<ExtensionEntryType, ExtensionType>(
