+++ shared/package.json
@@ -9,6 +9,7 @@
     "dev": "nodemon -e ts -w ./src -x npm run build"
   },
   "devDependencies": {
+    "@types/node": "^18.11.9",
     "rimraf": "3.0.2",
     "typescript": "4.8.4"
   },
