+++ src/kindhandlers/index.ts
@@ -8,7 +8,6 @@ import path from 'path';
 import {getStaticResourcePath} from '@redux/services';
 import {refMapperMatchesKind} from '@redux/services/resourceRefs';
 
-import {getSubfolders, readFiles} from '@utils/fileSystem';
 import {parseAllYamlDocuments} from '@utils/yaml';
 
 import EndpointSliceHandler from '@src/kindhandlers/EndpointSlice.handler';
@@ -20,6 +19,7 @@ import VolumeAttachmentHandler from '@src/kindhandlers/VolumeAttachment.handler'
 import {extractKindHandler} from '@src/kindhandlers/common/customObjectKindHandler';
 
 import {K8sResource, RefMapper, ResourceKindHandler} from '@monokle-desktop/shared/models';
+import {getSubfolders, readFiles} from '@monokle-desktop/shared/utils/fileSystem';
 
 import ClusterRoleHandler from './ClusterRole.handler';
 import ClusterRoleBindingHandler from './ClusterRoleBinding.handler';
